Student Record Search Project
Utilizes arrays, objects, and loops to create a program that searches a
database of students for information on an individual student and prints
that information.

User has the ability to manually end the loop using input and also
receives feedback based on whether a certain student exists in the
database or not. The program also accounts for duplicate names.